requirements.txt
async-generator==1.10
attrs==22.1.0
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.9.24
cfscrape==2.1.1
charset-normalizer==2.1.1
chromedriver-py==108.0.5359.71
colorama==0.4.6
colorthief==0.2.1
docopt==0.6.2
exceptiongroup==1.0.4
h11==0.14.0
idna==3.4
outcome==1.2.0
Pillow==9.3.0
pipreqs==0.4.11
PySocks==1.7.1
requests==2.28.1
selenium==4.6.1
selenium-stealth==1.0.6
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
trio==0.22.0
trio-websocket==0.9.2
urllib3==1.26.13
wsproto==1.2.0
yarg==0.1.9